Rename userdata to login and remove local dht
