+++ src/main/scala/login/LocatorInfo.scala
@@ -1,6 +1,6 @@
-package userData
+package login
 
-import userData.State.State
+import login.State.State
 
 case class LocatorInfo(
                         locator: String, // a string defined by user, to tell on which machine the user is currently active
