+++ src/main/scala/login/LogoutProcedure.scala
@@ -1,4 +1,4 @@
-package userData
+package login
 
 import dht.{DistributedDHT, GetPeerKey}
